extract deadline decoration, use in valinnantulosServlet if hakuOid defined
extract deadline decoration, use in valinnantulosServlet if hakuOid defined
