+++ valinta-tulos-service/src/test/scala/fi/vm/sade/valintatulosservice/local/ValinnantulosServletSpec.scala
@@ -32,8 +32,9 @@ class ValinnantulosServletSpec extends Specification with EmbeddedJettyContainer
 
   def foreach[R: AsResult](f: ((String, ValinnantulosService, SessionRepository)) => R): org.specs2.execute.Result = {
     val valinnantulosService = mock[ValinnantulosService]
+    val valintatulosService = mock[ValintatulosService]
     val sessionRepository = mock[SessionRepository]
-    val servlet = new ValinnantulosServlet(valinnantulosService, sessionRepository)(mock[Swagger])
+    val servlet = new ValinnantulosServlet(valinnantulosService, valintatulosService, sessionRepository)(mock[Swagger])
     ServletTest.withServlet(this, servlet, (uri: String) => AsResult(f((uri, valinnantulosService, sessionRepository))))
   }
